# api_v1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authorization",
        "description": "--------------------------------\nBasic configuration for swagger\n--------------------------------",
        "version": "3.0.0"
    },
    "paths": {
        "/auth/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register With New User",
                "description": "Register User",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Emailid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully registered"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login With User Credentials",
                "description": "Login User",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully lognin"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWTHeaderAuthentication": {
                "type": "apiKey",
                "description": "------------------------------\nJWT Authentication for swagger\n------------------------------",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}